+++ lib/lab42/empty_stream.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Lab42
+  class EmptyStream < Stream
+    def empty? = true
+    private
+    def initialize = nil
+  end
+end
+# SPDX-License-Identifier: AGPL-3.0-or-later