+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 * Complete rewrite to eliminate complexity and namespace pollution
 
+* License change to AGPL-3.0-or-later
+
+* Use Ruby 3.3.5 at least
+
+* Get rid of dependencies
 
 # 0.2.1 2020-02-02
 
@@ -10,3 +15,4 @@
 * Replaced `qed` and `ae` (thank you for your long service) with `speculate_about`
 
 * Fixed code coverage for CodeClimate
+<!--SPDX-License-Identifier: AGPL-3.0-or-later-->